Fix memory leak in xstatbar.c:draw
The call to XCopyArea in draw() generates one of two X events, either a GraphicsExpose or a NoExpose event. Since xstatbar does not have one of those traditional X event loops, those events accumulate in the event queue and lead to a memory leak. The easiest fix is to call XNextEvent after XCopyArea to consume the events it generated.

Thanks @farhaven